requirements.txt
-i https://pypi.org/simple
# AI Doctor - Core Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
python-dotenv==1.0.0
pydantic-settings==2.7.0

# AI and ML
groq==0.15.0

# Audio Processing
gtts==2.5.4
speechrecognition==3.13.0
pydub==0.25.1
pyaudio==0.2.14

# Image Processing
pillow==11.1.0

# Utilities
aiofiles==23.2.1
requests==2.32.3

# Optional: Keep Gradio for alternative interface
gradio==5.12.0
